+++ embrace/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.2.0
+
+* Added a method to manually log push notifications
+* Updated Embrace Android SDK to 5.17.1
+* Updated Embrace iOS SDK to 5.19.2
+
 # 1.1.0
 
 * Updated Embrace Android SDK to 5.16.0